+++ Examples/SwiftSampleApp/SwiftSampleApp/SampleCLogging.h
@@ -1,0 +1,7 @@
+#ifndef SampleCLogging_h
+#define SampleCLogging_h
+
+void SampleLogPrintf(const char *message);
+void SampleLogFprintfStderr(const char *message);
+
+#endif